move Parquet meta renderer to Perspective renderer
